Verify the job has been executed before trying to access to lastBuild
Since long ago we had detected that, on startup, the bot was missing the
status from a couple of our ci servers. After tracing the problem down,
it was caused by some jobs in those servers that never had been run,
hence their lastBuild was null.

This simply adds a check to ensure that some build has happened.